refactor: harden config handling and UI boundaries
Use atomic configuration and credential writes, extract preferences and metadata services, move dialog implementations out of headers, eliminate strict-build conversion warnings, and provide actionable configuration parse errors.

+++ src/core/mds_helpers.cpp
@@ -250,7 +250,7 @@ LayoutConfig expandedShotLayout(const LayoutConfig& config)
                     SignalSpec copy = sig;
                     copy.shot = shot;
                     copy.manualColor = false;
-                    copy.colorName = colorForIndex(expandedSignals.size());
+                    copy.colorName = colorForIndex(static_cast<int>(expandedSignals.size()));
                     expandedSignals.push_back(std::move(copy));
                 }
             }
